Ajout de la création d'un JSON + Methode POST pour TEST les intéractions avec le serveur

+++ BarTendr/BarTendr/Categorie.h
@@ -10,12 +10,12 @@
 
 @interface Categorie : NSObject {
     
-    int id_categorie;
+    long id_categorie;
     NSString *nom_categorie;
     
 }
 
-@property (nonatomic) int id_categorie;
+@property (nonatomic) long id_categorie;
 @property (nonatomic, retain) NSString *nom_categorie;
 
 @end
